refactor(merge-packages): move get-variable and rename merge-packages
