+++ packages/core/src/utils/__tests__/get-variable.tests.ts
@@ -1,0 +1,18 @@
+import getVariable from "../get-variable";
+
+describe("getVariable", () => {
+  it("should generate different variable names for different packages", () => {
+    expect(getVariable("@org/package", "mode")).not.toBe(
+      getVariable("org-package", "mode")
+    );
+    expect(getVariable("@org/package", "mode")).not.toBe(
+      getVariable("org.package", "mode")
+    );
+    expect(getVariable("org-package", "mode")).not.toBe(
+      getVariable("org.package", "mode")
+    );
+    expect(getVariable("@org/package", "html")).not.toBe(
+      getVariable("@org/package", "amp")
+    );
+  });
+});
