+++ packages/core/src/utils/get-variable.ts
@@ -1,0 +1,11 @@
+// Remove some characters present in the npm package name to
+// turn it into a variable name.
+export default (pkg: string, mode: string) => {
+  return (
+    pkg
+      .replace(/^@/, "")
+      .replace(/-/g, "_")
+      .replace(/\//g, "__")
+      .replace(/\./g, "___") + `_${mode}`
+  );
+};
