+++ packages/core/src/utils/merge-packages.ts
@@ -1,17 +1,6 @@
 import { Package } from "@frontity/types";
 import deepmerge from "deepmerge";
-
-// Remove some characters present in the npm package name to
-// turn it into a variable name.
-export const getVariable = (pkg: string, mode: string) => {
-  return (
-    pkg
-      .replace(/^@/, "")
-      .replace(/-/g, "_")
-      .replace(/\//g, "__")
-      .replace(/\./g, "___") + `_${mode}`
-  );
-};
+import getVariable from "./get-variable";
 
 type PackageFunction = ({
   libraries
@@ -21,7 +10,7 @@ type PackageFunction = ({
 
 // Merge all packages together in a single config that can be passed
 // to createStore.
-export const mergePackages = ({
+export default ({
   packages,
   state
 }: {
